# specs/pizza_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pizza Ordering API",
    "description": "Simple pizza ordering system for MCP transformation demo",
    "version": "1.0.0"
  },
  "paths": {
...
    "/orders": {
      "get": {
        "summary": "Get All Orders",
        "description": "Get all orders with proper schema",
        "operationId": "get_all_orders_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Pizza Order",
        "description": "Create a new pizza order with multiple items",
        "operationId": "create_pizza_order_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{order_id}/confirm": {
      "post": {
        "summary": "Confirm Order",
        "description": "Confirm order (business accepts) with notifications",
        "operationId": "confirm_order_orders__order_id__confirm_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/reject": {
      "post": {
        "summary": "Reject Order",
        "description": "Reject order (business declines)",
        "operationId": "reject_order_orders__order_id__reject_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/complete": {
      "post": {
        "summary": "Complete Order",
        "description": "Mark order as completed (ready for delivery) with notifications",
        "operationId": "complete_order_orders__order_id__complete_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderItem": {
        "properties": {
          "pizza": {
            "type": "string",
            "title": "Pizza"
          },
          "size": {
            "type": "string",
            "title": "Size"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          },
          "price": {
            "type": "number",
            "title": "Price",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "pizza",
          "size"
        ],
        "title": "OrderItem"
      },
      "OrderRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "type": "array",
            "title": "Items"
          },
          "customer_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Email"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "OrderRequest"
      },
      "OrderResponse": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "prep_time": {
            "type": "string",
            "title": "Prep Time"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "items",
          "total_amount",
          "status",
          "prep_time"
        ],
        "title": "OrderResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}